HRRN sigue dando mal pero por lo menos ordena ahora

+++ Kernel/include/recursos.h
@@ -22,7 +22,7 @@
     char* PUERTO_ESCUCHA;
     char* ALGORITMO_PLANIFICACION;
     int ESTIMACION_INICIAL;
-    int HRRN_ALFA;
+    double HRRN_ALFA;
     int GRADO_MAX_MULTIPROGRAMACION;
     t_list* RECURSOS;//es una lista de char
     t_list* INSTANCIAS_RECURSOS;// es una lista de int
